src/requirements.txt
accelerate
datasets
hydra-core
safetensors
tensorboard
torch
transformers
flask